Revert "Merge remote-tracking branch 'origin/master'"
This reverts commit b73bfab9719e1951699fc52507cbbf6101c357dc, reversing
changes made to d5620b4603211a22f801bfb1d8504cbc97fcdecd.
