+++ dap/BESDapFunctionResponseCache.h
@@ -27,7 +27,6 @@
 #define _bes_dap_function_response_cache_h
 
 #include <string>
-#include <mutex>
 
 #include "BESFileLockingCache.h"
 
@@ -73,17 +72,22 @@ class BaseTypeFactory;
 class BESDapFunctionResponseCache: public BESFileLockingCache {
 private:
     static bool d_enabled;
-    static std::once_flag d_initialize;
-#if 0
     static BESDapFunctionResponseCache *d_instance;
-#endif
 
+//    /**
+//     * Called by atexit()
+//     */
+//    static void delete_instance() {
+//        delete d_instance;
+//        d_instance = 0;
+//    }
 
     /** @name Suppressed constructors
      *  Do not use.
      */
     ///@{
-    BESDapFunctionResponseCache() = default;
+    BESDapFunctionResponseCache();
+    // BESDapFunctionResponseCache(const BESDapFunctionResponseCache &src);
     ///@}
 
     bool is_valid(const std::string &cache_file_name, const std::string &dataset);
@@ -101,7 +105,6 @@ private:
     friend class FunctionResponseCacheTest;
     friend class StoredResultTest;
 
-#if 0
 protected:
     /**
      * @brief Protected constructor that takes as arguments keys to the cache directory,
@@ -121,7 +124,6 @@ protected:
         BESFileLockingCache(cache_dir, prefix, size)
     {
     }
-#endif
 
 public:
     static const std::string PATH_KEY;
@@ -131,11 +133,8 @@ public:
     BESDapFunctionResponseCache(const BESDapFunctionResponseCache&) = delete;
     BESDapFunctionResponseCache& operator=(const BESDapFunctionResponseCache&) = delete;
 
-#if 0
     static BESDapFunctionResponseCache *get_instance(const std::string &cache_dir, const std::string &prefix,
                                                      unsigned long long size);
-#endif
-
     static BESDapFunctionResponseCache *get_instance();
 
     ~BESDapFunctionResponseCache() override = default;
