+++ dap/unit-tests/.gitignore
@@ -10,4 +10,3 @@ mds_inventory
 /pathinfo_files/
 /ShowPathInfoTest
 /DapUtilsTest
-/FunctionResponseCacheTest2
