+++ dispatch/unit-tests/.gitignore
@@ -13,4 +13,3 @@ uncompressT
 /RequestTimerTest
 /BESFileLockingCacheTest
 /FileCacheTest
-/uncompressT2
