+++ modules/csv_handler/tests/bes.conf.in
@@ -35,10 +35,6 @@ BES.UncompressCache.dir=/tmp
 BES.UncompressCache.prefix=uncompress_cache
 BES.UncompressCache.size=500
 
-# DAP.FunctionResponseCache.path=response_cache
-# DAP.FunctionResponseCache.prefix=/rc_
-# DAP.FunctionResponseCache.size=500
-
 BES.Container.Persistence=strict
 
 BES.Memory.GlobalArea.EmergencyPoolSize=1
